+++ base/src/test/java/com/plugatar/xteps/base/reporter/DefaultStepReporterTest.java
@@ -16,7 +16,7 @@
 package com.plugatar.xteps.base.reporter;
 
 import com.plugatar.xteps.base.ExceptionHandler;
-import com.plugatar.xteps.base.SafeACContainer;
+import com.plugatar.xteps.base.HookContainer;
 import com.plugatar.xteps.base.StepListener;
 import com.plugatar.xteps.base.StepReporter;
 import org.junit.jupiter.api.Test;
@@ -61,7 +61,7 @@ final class DefaultStepReporterTest {
         final StepListener stepListener1 = mock(StepListener.class);
         final StepListener stepListener2 = mock(StepListener.class);
         final StepReporter reporter = new DefaultStepReporter(new StepListener[]{stepListener1, stepListener2});
-        final SafeACContainer container = mock(SafeACContainer.class);
+        final HookContainer container = mock(HookContainer.class);
         final ExceptionHandler handler = mock(ExceptionHandler.class);
         final Object[] contexts = new Object[]{};
         final Object expectedResult = new Object();
@@ -80,7 +80,7 @@ final class DefaultStepReporterTest {
         final StepListener stepListener1 = mock(StepListener.class);
         final StepListener stepListener2 = mock(StepListener.class);
         final StepReporter reporter = new DefaultStepReporter(new StepListener[]{stepListener1, stepListener2});
-        final SafeACContainer container = mock(SafeACContainer.class);
+        final HookContainer container = mock(HookContainer.class);
         final ExceptionHandler handler = mock(ExceptionHandler.class);
         final Object[] contexts = new Object[]{};
         final RuntimeException expectedException = new RuntimeException();