+++ base/src/test/java/com/plugatar/xteps/base/reporter/FakeStepReporterTest.java
@@ -16,7 +16,7 @@
 package com.plugatar.xteps.base.reporter;
 
 import com.plugatar.xteps.base.ExceptionHandler;
-import com.plugatar.xteps.base.SafeACContainer;
+import com.plugatar.xteps.base.HookContainer;
 import com.plugatar.xteps.base.StepReporter;
 import org.junit.jupiter.api.Test;
 
@@ -35,7 +35,7 @@ final class FakeStepReporterTest {
     @Test
     void reportStepWithoutException() {
         final StepReporter reporter = new FakeStepReporter();
-        final SafeACContainer container = mock(SafeACContainer.class);
+        final HookContainer container = mock(HookContainer.class);
         final ExceptionHandler handler = mock(ExceptionHandler.class);
         final Object[] contexts = new Object[]{};
         final Object expectedResult = new Object();
@@ -50,7 +50,7 @@ final class FakeStepReporterTest {
     @Test
     void reportStepWithException() {
         final StepReporter reporter = new FakeStepReporter();
-        final SafeACContainer container = mock(SafeACContainer.class);
+        final HookContainer container = mock(HookContainer.class);
         final ExceptionHandler handler = mock(ExceptionHandler.class);
         final Object[] contexts = new Object[]{};
         final RuntimeException expectedException = new RuntimeException();
@@ -60,7 +60,7 @@ final class FakeStepReporterTest {
                 throw expectedException;
             });
         }).isSameAs(expectedException);
-        verify(container).close(same(expectedException));
+        verify(container).callHooks(same(expectedException));
         verify(handler).handle(same(expectedException));
     }
 }